server/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
filelock==3.13.3
fonttools==4.54.1
idna==3.8
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pipenv==2023.12.1
platformdirs==4.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rake-nltk==1.0.6
regex==2024.9.11
requests==2.32.3
setuptools==69.2.0
six==1.16.0
soupsieve==2.6
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.2
virtualenv==20.25.1
